move GetProfileForSysObjectID
