+++ pkg/collector/corechecks/snmp/internal/checkconfig/testing_utils.go
@@ -6,33 +6,8 @@
 package checkconfig
 
 import (
-	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/internal/profile"
-	"os"
-	"path/filepath"
-
-	"github.com/DataDog/datadog-agent/pkg/config"
-
 	"github.com/DataDog/datadog-agent/pkg/networkdevice/profile/profiledefinition"
 )
-
-// SetConfdPathAndCleanProfiles is used for testing only
-func SetConfdPathAndCleanProfiles() {
-	profile.SetGlobalProfileConfigMap(nil) // make sure from the new confd path will be reloaded
-	file, _ := filepath.Abs(filepath.Join(".", "test", "conf.d"))
-	if !pathExists(file) {
-		file, _ = filepath.Abs(filepath.Join("..", "test", "conf.d"))
-	}
-	if !pathExists(file) {
-		file, _ = filepath.Abs(filepath.Join(".", "internal", "test", "conf.d"))
-	}
-	config.Datadog.Set("confd_path", file)
-}
-
-// pathExists returns true if the given path exists
-func pathExists(path string) bool {
-	_, err := os.Stat(path)
-	return !os.IsNotExist(err)
-}
 
 // copyProfileDefinition copies a profile, it's used for testing
 func copyProfileDefinition(profileDef profiledefinition.ProfileDefinition) profiledefinition.ProfileDefinition {
