+++ pkg/collector/corechecks/snmp/internal/profile/testing_utils.go
@@ -1,7 +1,6 @@
-package checkconfig
+package profile
 
 import (
-	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/internal/profile"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"os"
 	"path/filepath"
@@ -9,7 +8,7 @@ import (
 
 // SetConfdPathAndCleanProfiles is used for testing only
 func SetConfdPathAndCleanProfiles() {
-	profile.SetGlobalProfileConfigMap(nil) // make sure from the new confd path will be reloaded
+	SetGlobalProfileConfigMap(nil) // make sure from the new confd path will be reloaded
 	file, _ := filepath.Abs(filepath.Join(".", "test", "conf.d"))
 	if !pathExists(file) {
 		file, _ = filepath.Abs(filepath.Join("..", "test", "conf.d"))
