+++ pkg/collector/corechecks/snmp/snmp_test.go
@@ -10,6 +10,7 @@ import (
 	"encoding/json"
 	"fmt"
 	common2 "github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/internal/common"
+	"github.com/DataDog/datadog-agent/pkg/collector/corechecks/snmp/internal/profile"
 	"testing"
 	"time"
 
@@ -59,7 +60,7 @@ func createDeps(t *testing.T) deps {
 
 func Test_Run_simpleCase(t *testing.T) {
 	deps := createDeps(t)
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -338,7 +339,7 @@ tags:
 
 func Test_Run_customIfSpeed(t *testing.T) {
 	deps := createDeps(t)
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -476,7 +477,7 @@ metrics:
 }
 
 func TestSupportedMetricTypes(t *testing.T) {
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -554,7 +555,7 @@ func TestProfile(t *testing.T) {
 	deps := createDeps(t)
 	senderManager := aggregator.InitAndStartAgentDemultiplexer(deps.Log, deps.Forwarder, demuxOpts(), "")
 
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
@@ -939,7 +940,7 @@ profiles:
 }
 
 func TestServiceCheckFailures(t *testing.T) {
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -975,7 +976,7 @@ community_string: public
 }
 
 func TestCheckID(t *testing.T) {
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	check1 := snmpFactory()
 	check2 := snmpFactory()
 	check3 := snmpFactory()
@@ -1186,7 +1187,7 @@ func TestCheck_Run(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			checkconfig.SetConfdPathAndCleanProfiles()
+			profile.SetConfdPathAndCleanProfiles()
 			sess := session.CreateMockSession()
 			sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 				return sess, nil
@@ -1236,7 +1237,7 @@ namespace: '%s'
 }
 
 func TestCheck_Run_sessionCloseError(t *testing.T) {
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
@@ -1284,7 +1285,7 @@ func TestReportDeviceMetadataEvenOnProfileError(t *testing.T) {
 
 	deps := createDeps(t)
 	senderManager := aggregator.InitAndStartAgentDemultiplexer(deps.Log, deps.Forwarder, demuxOpts(), "")
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
@@ -1590,7 +1591,7 @@ func TestReportDeviceMetadataWithFetchError(t *testing.T) {
 	deps := createDeps(t)
 	senderManager := aggregator.InitAndStartAgentDemultiplexer(deps.Log, deps.Forwarder, demuxOpts(), "")
 
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
@@ -1696,7 +1697,7 @@ tags:
 func TestDiscovery(t *testing.T) {
 	deps := createDeps(t)
 	timeNow = common2.MockTimeNow
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -2035,7 +2036,7 @@ metric_tags:
 
 func TestDiscovery_CheckError(t *testing.T) {
 	deps := createDeps(t)
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
@@ -2112,7 +2113,7 @@ func TestDeviceIDAsHostname(t *testing.T) {
 	deps := createDeps(t)
 	cache.Cache.Delete(cache.BuildAgentKey("hostname")) // clean existing hostname cache
 
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -2303,7 +2304,7 @@ func TestDiscoveryDeviceIDAsHostname(t *testing.T) {
 	deps := createDeps(t)
 	cache.Cache.Delete(cache.BuildAgentKey("hostname")) // clean existing hostname cache
 	timeNow = common2.MockTimeNow
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
@@ -2508,7 +2509,7 @@ metrics:
 
 func TestCheckCancel(t *testing.T) {
 	deps := createDeps(t)
-	checkconfig.SetConfdPathAndCleanProfiles()
+	profile.SetConfdPathAndCleanProfiles()
 	sess := session.CreateMockSession()
 	sessionFactory := func(*checkconfig.CheckConfig) (session.Session, error) {
 		return sess, nil
